+++ NEWS.md
@@ -10,6 +10,10 @@ Other improvements:
 - `mp4` files are better encoded and should be playable on more devices
 - When `verbose`, now `ffmpeg` displays its messages
 
+Bug fixes:
+- Do not require movie file when `--only-render-movie` and `--frame-name-format`
+  are passed.
+
 ## Version 0.2.0 (5 April 2021)
 
 New features added: